requirements.txt
bs4
requests
httpx
streamlit
streamlit-scrollable-textbox
uvicorn
fastapi
openai
python-dotenv